renamed mason to faber, changed library name, updated source with test cases
