+++ visualiser/index.html
@@ -16,7 +16,7 @@
     }
 
     .html,
-    .mason {
+    .faber {
       height: 300px;
       width: 300px;
       border: 1px solid #ffff00;
@@ -29,17 +29,17 @@
 </head>
 
 <body>
-  <h1>Do not fear, Mason is here!</h1>
+  <h1>Do not fear, faber is here!</h1>
 
   <div class="container">
     <div class="html">
     </div>
 
-    <div class="mason">
+    <div class="faber">
     </div>
   </div>
   <script src="../dist/main.js"></script>
-  <script src="./css-mason-compartor.js"></script>
+  <script src="./css-faber-compartor.js"></script>
   <script>
     getHTMLCSSInlined();
   </script>
